+++ src/Config.js
@@ -18,9 +18,9 @@ const { isFunction, isObject } = require("lodash");
 const CONFIG_FILE = "mocks-server.config.js";
 
 class Config {
-  constructor(programmaticConfig = {}, { addAlert }) {
+  constructor({ addAlert }, programmaticConfig = {}) {
     this._coreOptions = {};
-    this._onError = addAlert;
+    this._addAlert = addAlert;
     this._options = {};
     this._getCoreOptions(programmaticConfig);
     this._getOptions(programmaticConfig.options);
@@ -32,7 +32,7 @@ class Config {
   }
 
   _catchFileConfigError(error) {
-    this._onError("file", "Error in configuration file", error);
+    this._addAlert("file", "Error in configuration file", error);
     return Promise.resolve({});
   }