+++ test/unit/plugins/Plugins.spec.js
@@ -22,6 +22,7 @@ const pluginsQuantity = (method, quantity) => {
 };
 
 describe("Plugins", () => {
+  let callbacks;
   let sandbox;
   let coreMocks;
   let coreInstance;
@@ -32,6 +33,10 @@ describe("Plugins", () => {
 
   beforeEach(async () => {
     sandbox = sinon.createSandbox();
+    callbacks = {
+      addAlert: sandbox.stub(),
+      removeAlerts: sandbox.stub(),
+    };
     sandbox.stub(tracer, "verbose");
     sandbox.stub(tracer, "debug");
     sandbox.stub(tracer, "error");
@@ -54,7 +59,7 @@ describe("Plugins", () => {
   describe("register method", () => {
     const METHOD = "Register";
     it("should do nothing if there are no plugins to register", async () => {
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -62,7 +67,7 @@ describe("Plugins", () => {
     it("should register object plugins", async () => {
       const fooPlugin = {};
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
@@ -72,19 +77,38 @@ describe("Plugins", () => {
         register: () => {},
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
-    it("should register object plugins with register method passing to it the core itself", async () => {
+    it("should register object plugins with register method passing to them the core itself", async () => {
       const fooPlugin = {
         register: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(fooPlugin.register.calledWith(coreInstance)).toEqual(true);
+    });
+
+    it("should register object plugins with register method passing to them custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      const fooPlugin = {
+        register: (coreInstance, methods) => {
+          methods.loadMocks();
+          methods.addAlert("foo", "Foo message");
+          methods.removeAlerts();
+        },
+      };
+      configInstance.coreOptions.plugins = [fooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should not register object plugins with register method throwing an error", async () => {
@@ -94,7 +118,7 @@ describe("Plugins", () => {
         },
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -102,18 +126,22 @@ describe("Plugins", () => {
     it("should not register strings as plugins", async () => {
       expect.assertions(2);
       configInstance.coreOptions.plugins = ["foo"];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
     it("should not register booleans as plugins", async () => {
       expect.assertions(2);
       configInstance.coreOptions.plugins = [true];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
@@ -121,11 +149,29 @@ describe("Plugins", () => {
       expect.assertions(3);
       const fooPlugin = sinon.spy();
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(fooPlugin.calledWith(coreInstance)).toEqual(true);
       expect(fooPlugin.callCount).toEqual(1);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
+    });
+
+    it("should register function plugins executing them passing custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      const fooPlugin = (coreInstance, methods) => {
+        methods.loadMocks();
+        methods.addAlert("foo", "Foo message");
+        methods.removeAlerts();
+        return {};
+      };
+      configInstance.coreOptions.plugins = [fooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should register function plugins returning a register method", async () => {
@@ -135,7 +181,7 @@ describe("Plugins", () => {
         register: spy,
       });
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(spy.calledWith(coreInstance)).toEqual(true);
       expect(spy.callCount).toEqual(1);
@@ -149,7 +195,7 @@ describe("Plugins", () => {
         },
       });
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -166,11 +212,31 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(receivedCore).toEqual(coreInstance);
       expect(instantiated).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
+    });
+
+    it("should register class plugins, instantiating them passing custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      class FooPlugin {
+        constructor(core, methods) {
+          methods.loadMocks();
+          methods.addAlert("foo", "Foo message");
+          methods.removeAlerts();
+          return {};
+        }
+      }
+      configInstance.coreOptions.plugins = [FooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should not register class plugins if class throw an error when being created", async () => {
@@ -181,9 +247,11 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
@@ -201,7 +269,7 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(receivedCore).toEqual(coreInstance);
       expect(instantiated).toEqual(true);
@@ -219,13 +287,13 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
     it("should trace the total number of registered plugins", async () => {
-      expect.assertions(2);
+      expect.assertions(4);
       class FooPlugin {
         constructor() {
           throw new Error();
@@ -241,9 +309,17 @@ describe("Plugins", () => {
         "foo",
         { foo: "foo" },
       ];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:4", 'Error registering plugin "4"')).toEqual(
+        true
+      );
+      expect(callbacks.addAlert.calledWith("register:5", 'Error registering plugin "5"')).toEqual(
+        true
+      );
+      expect(callbacks.addAlert.calledWith("register:6", 'Error registering plugin "6"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 4))).toEqual(true);
     });
   });
@@ -252,7 +328,7 @@ describe("Plugins", () => {
     const METHOD = "Initializat";
     it("should do nothing if there are no plugins to register", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 0))).toEqual(true);
@@ -264,7 +340,7 @@ describe("Plugins", () => {
         init: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(fooPlugin.init.callCount).toEqual(1);
@@ -277,7 +353,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.debug.calledWith('Initializing plugin "foo-plugin"')).toEqual(true);
@@ -292,14 +368,14 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
-    it("should catch init method errors", async () => {
-      expect.assertions(1);
+    it("should catch init method errors and notify alerts", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         init: () => {
           throw new Error();
@@ -312,9 +388,13 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
+      expect(callbacks.removeAlerts.calledWith("init")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("init:1", 'Error initializating plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
@@ -334,7 +414,7 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
@@ -350,7 +430,7 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
@@ -362,7 +442,7 @@ describe("Plugins", () => {
 
     it("should do nothing if there are no plugins to register", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
@@ -374,7 +454,7 @@ describe("Plugins", () => {
         start: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(fooPlugin.start.callCount).toEqual(1);
@@ -387,7 +467,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.debug.calledWith('Starting plugin "foo-plugin"')).toEqual(true);
@@ -402,14 +482,14 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
-    it("should catch start method errors", async () => {
-      expect.assertions(1);
+    it("should catch start method errors and notify alert", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         start: () => {
           throw new Error();
@@ -422,9 +502,11 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
+      expect(callbacks.removeAlerts.calledWith("start")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("start:1", 'Error starting plugin "1"')).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
@@ -444,7 +526,7 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -460,7 +542,7 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -472,7 +554,7 @@ describe("Plugins", () => {
 
     it("should do nothing if there are no plugins to stop", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
@@ -484,7 +566,7 @@ describe("Plugins", () => {
         stop: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(fooPlugin.stop.callCount).toEqual(1);
@@ -497,7 +579,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.debug.calledWith('Stopping plugin "foo-plugin"')).toEqual(true);
@@ -512,14 +594,14 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
-    it("should catch stop method errors", async () => {
-      expect.assertions(1);
+    it("should catch stop method errors and notify alert", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         stop: () => {
           throw new Error();
@@ -532,9 +614,11 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
+      expect(callbacks.removeAlerts.calledWith("stop")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("stop:1", 'Error stopping plugin "1"')).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
@@ -554,7 +638,7 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -570,7 +654,7 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);