+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Add alerts handler, allowing plugins and other internal pieces to inform about errors or warnings, and let plugins access to the list, so they can display it, for example. (#102)
 ### Changed
 ### Fixed
+- fix: Avoid errors when calling to fixtures and behaviors getters if they are not already initialized
 ### Removed
 
 ## [1.5.2] - 2020-12-21